fix two issues with gcc 4.4
svn path=/trunk/gtk/; revision=56983

+++ src/YGBarGraph.cc
@@ -38,7 +38,7 @@ public:
 
 	static GdkColor ycolorToGdk (const YColor &ycolor)
 	{
-		GdkColor color = { 0, ycolor.red() << 8, ycolor.green() << 8, ycolor.blue() << 8 };
+		GdkColor color = { 0, guint16(ycolor.red() << 8), guint16(ycolor.green() << 8), guint16(ycolor.blue() << 8) };
 		return color;
 	}
